+++ packages/app/src/features/dialogs/e-mode/views/EModeView.tsx
@@ -11,7 +11,7 @@ import { PageStatus } from '@/features/dialogs/common/types'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
 import { Link } from '@/ui/atoms/link/Link'
 import { links } from '@/ui/constants/links'
-import { Alert } from '@/ui/molecules/new/alert/Alert'
+import { Alert } from '@/ui/molecules/alert/Alert'
 import { RiskAcknowledgement } from '@/ui/organisms/risk-acknowledgement/RiskAcknowledgement'
 import { CategoriesGrid } from '../components/CategoriesGrid'
 import { EModeCategoryTile } from '../components/EModeCategoryTile'
